reorganize rpc
