+++ src/eth/mod.rs
@@ -1,5 +1,4 @@
 pub mod codegen;
-pub mod decode;
 pub mod executor;
 pub mod external_rpc;
 pub mod follower;
