+++ src/eth/rpc/middleware/mod.rs
@@ -1,0 +1,9 @@
+mod decode;
+mod http_middleware;
+pub mod multicall;
+mod rpc_middleware;
+
+pub use decode::decode_input_arguments;
+pub use http_middleware::Authentication;
+pub use http_middleware::RpcHttpMiddleware;
+pub use rpc_middleware::RpcMiddleware;
