+++ src/eth/rpc/middleware/rpc_middleware.rs
@@ -31,10 +31,10 @@ use crate::eth::codegen::ContractName;
 use crate::eth::codegen::SoliditySignature;
 use crate::eth::rpc::RpcClientApp;
 use crate::eth::rpc::RpcError;
-use crate::eth::rpc::multicall::MulticallInfo;
+use crate::eth::rpc::middleware::multicall::MulticallInfo;
 use crate::eth::rpc::next_rpc_param;
 use crate::eth::rpc::parse_rpc_rlp;
-use crate::eth::rpc::rpc_parser::RpcExtensionsExt;
+use crate::eth::rpc::parser::RpcExtensionsExt;
 use crate::eth::types::Address;
 use crate::eth::types::Bytes;
 use crate::eth::types::CallInput;
