+++ src/eth/rpc/parser.rs
@@ -5,9 +5,9 @@ use jsonrpsee::types::ParamsSequence;
 use rlp::Decodable;
 use tracing::Span;
 
-use super::rpc_http_middleware::Authentication;
+use super::middleware::Authentication;
+use crate::eth::rpc::RpcClientApp;
 use crate::eth::rpc::RpcError;
-use crate::eth::rpc::rpc_client_app::RpcClientApp;
 use crate::ext::type_basename;
 use crate::infra::tracing::EnteredWrap;
