+++ src/eth/rpc/types/mod.rs
@@ -2,6 +2,7 @@ mod block_filter;
 mod error;
 mod log_filter;
 mod log_filter_input;
+mod rpc_client_app;
 mod timestamp_filter;
 
 pub use block_filter::BlockFilter;
@@ -10,5 +11,6 @@ pub use error::RpcError;
 pub use log_filter::LogFilter;
 pub use log_filter_input::LogFilterInput;
 pub use log_filter_input::LogFilterInputTopic;
+pub use rpc_client_app::RpcClientApp;
 pub use timestamp_filter::BlockTimestampFilter;
 pub use timestamp_filter::BlockTimestampSeekMode;
