+++ src/eth/types/block/mod.rs
@@ -1,12 +1,9 @@
-mod account;
 #[allow(clippy::module_inception)]
 mod block;
 mod block_header;
 mod pending_block;
 mod pending_block_header;
 
-pub use account::Account;
-pub use account::test_accounts;
 pub use block::Block;
 pub use block_header::BlockHeader;
 pub use pending_block::PendingBlock;
