+++ src/eth/types/mod.rs
@@ -4,12 +4,10 @@ pub mod execution_kind;
 pub mod external;
 pub mod primitives;
 pub mod transaction;
-pub use block::Account;
 pub use block::Block;
 pub use block::BlockHeader;
 pub use block::PendingBlock;
 pub use block::PendingBlockHeader;
-pub use block::test_accounts;
 pub use error::DecodeInputError;
 pub use error::ErrorCode;
 pub use error::StateError;
@@ -22,6 +20,7 @@ pub use external::ExternalBlockWithReceipts;
 pub use external::ExternalReceipt;
 pub use external::ExternalReceipts;
 pub use external::ExternalTransaction;
+pub use primitives::Account;
 pub use primitives::Address;
 pub use primitives::BlockNumber;
 pub use primitives::Bytes;
